+++ Releases/Utility Mod/Changelog.md
@@ -2,6 +2,7 @@
 - Allow TankBro's tank to spawn when enemy spawns are disabled.
 - Added fix for reversed and slow scroll wheel in mod manager UI on Linux.
 - Updated Swap Bros integration to use new API.
+- Fix level edits not replaying when dying and restarting on a level.
 
 ## v2.4.1
 - Added option to go to custom campaigns from right click menu.